+++ e2e/jasmine-async/__tests__/pending_in_promise.test.js
@@ -1,0 +1,15 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @jest-environment jsdom
+ */
+
+'use strict';
+
+it('skips a test inside a promise', () =>
+  new Promise(() => {
+    pending('skipped a test inside a promise');
+  }));